refactoring code in interfacerenderer and robotrenderer
