+++ src/main/java/biblioteket/roborally/game/GameLoop.java
@@ -1,8 +1,6 @@
 package biblioteket.roborally.game;
 
 import biblioteket.roborally.actors.IPlayer;
-import biblioteket.roborally.actors.Player;
-import biblioteket.roborally.actors.RobotRenderer;
 import biblioteket.roborally.board.DirVector;
 import biblioteket.roborally.board.Direction;
 import biblioteket.roborally.board.IBoard;
@@ -16,7 +14,7 @@ import biblioteket.roborally.programcards.CardDeck;
 import biblioteket.roborally.programcards.ICard;
 import biblioteket.roborally.programcards.ICardDeck;
 import biblioteket.roborally.programcards.ReverseCardComparator;
-import biblioteket.roborally.userinterface.InterfaceRenderer;
+import biblioteket.roborally.actors.InterfaceRenderer;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
 import com.badlogic.gdx.InputAdapter;
