+++ src/main/java/biblioteket/roborally/game/GameScreen.java
@@ -4,8 +4,7 @@ import biblioteket.roborally.actors.*;
 import biblioteket.roborally.board.Board;
 import biblioteket.roborally.board.IBoard;
 import biblioteket.roborally.elements.ArchiveMarkerElement;
-import biblioteket.roborally.userinterface.InterfaceRenderer;
-import com.badlogic.gdx.Game;
+import biblioteket.roborally.actors.InterfaceRenderer;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
@@ -53,14 +52,14 @@ public class GameScreen implements Screen {
 
         this.players = new ArrayList<>();
 
-        for (int i = 1; i <= 2; i++) {
+        for (int i = 1; i <= 4; i++) {
             TiledMapTileLayer.Cell playerCell = new TiledMapTileLayer.Cell().setTile(new StaticTiledMapTile(playerTextureSplit[0][0]));
             Player player = new Player(board, playerCell, new InterfaceRenderer(), robotRenderer);
             players.add(player);
             ArchiveMarkerElement archiveMarker = board.getArchiveMarker(i);
             IRobot robot = new Robot(archiveMarker);
             player.setRobot(robot);
-            player.setName("Player " + Integer.toString(i));
+            player.setName("Player " + i);
             board.getPlayerLayer().setCell(player.getRobot().getPosition().getX(), player.getRobot().getPosition().getY(), playerCell);
         }
 
@@ -79,6 +78,7 @@ public class GameScreen implements Screen {
         Gdx.gl.glClearColor(0, 0, 0, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT); // Clears main menu screen
 
+        // Render interface of current player
         gameLoop.getCurrentPlayer().getInterfaceRenderer().renderInterface(board);
 
         // Render robot movement
